requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
PyMySQL
slowapi
